src/main/java/com/wgl/sell/service/impl/CategoryServiceImpl.java
package com.wgl.sell.service.impl;

import com.wgl.sell.dao.ProductCategoryDao;
import com.wgl.sell.entity.ProductCategory;
import com.wgl.sell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryDao productCategoryDao;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return productCategoryDao.findById(categoryId).get();
    }

    @Override
    public List<ProductCategory> findAll() {
        return productCategoryDao.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryList) {
        return productCategoryDao.findByCategoryTypeIn(categoryList);
    }

    @Override
    public ProductCategory Save(ProductCategory productCategory) {
        return productCategoryDao.save(productCategory);
    }
}
